Fixed FEDN-107
